Fix polars ancestor_id col dtype: cast to Int64 instead of UInt64
The polars ancestor_list parser was casting to UInt64, which when
combined with fill_null(pl.col("id")) (Int64) caused polars to promote
to Float64. Fix by casting to Int64 to match the id column type.
Revert caller-side .astype(np.int64) workarounds that are no longer
needed.

+++ hstrat/_auxiliary_lib/_alifestd_calc_mrca_id_vector_asexual.py
@@ -77,7 +77,7 @@ def alifestd_calc_mrca_id_vector_asexual(
     if target_id >= len(phylogeny_df):
         raise ValueError(f"{target_id=} out of bounds")
 
-    ancestor_ids = phylogeny_df["ancestor_id"].to_numpy().astype(np.int64)
+    ancestor_ids = phylogeny_df["ancestor_id"].to_numpy()
     assert np.all(
         phylogeny_df["id"].to_numpy() == np.arange(len(phylogeny_df))
     )
